Refactor constructors
